requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
opencv-python-headless>=4.10.0
mediapipe==0.10.14
numpy>=1.26.0
pillow>=10.0.0
pillow-heif>=0.18.0
pydantic>=2.8.0
pyyaml>=6.0.0
